[bug] fix fade time

+++ number-selector/Assets/Scripts/Config.cs
@@ -1,4 +1,4 @@
 public static class Config
 {
-    public static int FADE_TIME_IN_SECONDS = 2;
+    public static float FADE_TIME_IN_SECONDS = 2f;
 }
